directory name change
