--- src/main/java/note/example/NoteApplication.java
@@ -1,13 +1,0 @@
-package note.example;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NoteApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NoteApplication.class, args);
-	}
-
-}
