--- src/main/java/note/example/constant/ConstantValue.java
@@ -1,5 +1,0 @@
-package note.example.constant;
-
-public class ConstantValue {
-    public static final String OPERATION_SUCCESSFUL = "Operation successful";
-}
