--- src/main/java/note/example/controller/UserController.java
@@ -1,13 +1,0 @@
-package note.example.controller;
-
-import note.example.dto.response.CommonResponse;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController {
-    @GetMapping("/test")
-    public CommonResponse test() {
-        return null;
-    }
-}
