--- src/main/java/note/example/dto/response/CommonResponse.java
@@ -1,11 +1,0 @@
-package note.example.dto.response;
-
-import lombok.Data;
-import lombok.experimental.SuperBuilder;
-
-@SuperBuilder
-@Data
-public class CommonResponse {
-    protected Integer code;
-    protected String message;
-}
