--- src/main/java/note/example/exception/ExceptionHandling.java
@@ -1,36 +1,0 @@
-package note.example.exception;
-
-import note.example.dto.response.CommonResponse;
-import note.example.utility.Util;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class ExceptionHandling {
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity<Object> handleGlobalException(Exception ex) {
-        return Util.form(
-                buildResponse(
-                        HttpStatus.BAD_REQUEST.value(),
-                        ex.getMessage()),
-                HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(NotFoundException.class)
-    public ResponseEntity<Object> handleNotFoundException(Exception ex) {
-        return Util.form(
-                buildResponse(
-                        HttpStatus.NOT_FOUND.value(),
-                        ex.getMessage()),
-                HttpStatus.NOT_FOUND);
-    }
-
-    public CommonResponse buildResponse(Integer code, String message) {
-        return CommonResponse.builder()
-                .code(code)
-                .message(message)
-                .build();
-    }
-}
