--- src/main/java/note/example/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package note.example.exception;
-
-public class NotFoundException extends RuntimeException{
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
