--- src/main/java/note/example/service/AuthService.java
@@ -1,13 +1,0 @@
-package note.example.service;
-
-import note.example.dto.request.LoginRequest;
-import note.example.dto.request.SignupRequest;
-import note.example.dto.response.CommonResponse;
-import note.example.dto.response.LoginResponse;
-import note.example.dto.response.SignupResponse;
-
-public interface AuthService {
-    SignupResponse userSignUp(SignupRequest signupRequest);
-
-    LoginResponse userLogin(LoginRequest loginRequest);
-}
