--- src/main/java/note/example/service/UserService.java
@@ -1,4 +1,0 @@
-package note.example.service;
-
-public interface UserService {
-}
