--- src/main/java/note/example/service/impl/UserServiceImpl.java
@@ -1,5 +1,0 @@
-package note.example.service.impl;
-
-public class UserServiceImpl {
-
-}
