+++ src/main/java/note/share/NoteApplication.java
@@ -1,0 +1,13 @@
+package note.share;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NoteApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NoteApplication.class, args);
+	}
+
+}
