+++ src/main/java/note/share/constant/ConstantValue.java
@@ -1,0 +1,5 @@
+package note.share.constant;
+
+public class ConstantValue {
+    public static final String OPERATION_SUCCESSFUL = "Operation successful";
+}
