+++ src/main/java/note/share/controller/UserController.java
@@ -1,0 +1,13 @@
+package note.share.controller;
+
+import note.share.dto.response.CommonResponse;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+    @GetMapping("/test")
+    public CommonResponse test() {
+        return null;
+    }
+}
