+++ src/main/java/note/share/dto/response/CommonResponse.java
@@ -1,0 +1,11 @@
+package note.share.dto.response;
+
+import lombok.Data;
+import lombok.experimental.SuperBuilder;
+
+@SuperBuilder
+@Data
+public class CommonResponse {
+    protected Integer code;
+    protected String message;
+}
