+++ src/main/java/note/share/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package note.share.exception;
+
+public class NotFoundException extends RuntimeException{
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
