+++ src/main/java/note/share/model/BasicAttribute.java
@@ -1,0 +1,26 @@
+package note.share.model;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.Instant;
+
+@Getter
+@Setter
+@MappedSuperclass
+public class BasicAttribute {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    public Long id;
+
+    @CreationTimestamp
+    @Column(name = "created_at", updatable = false, nullable = false)
+    public Instant createdAt;
+
+    @UpdateTimestamp
+    @Column(name = "last_modified_at", updatable = true, nullable = false)
+    public Instant lastModifiedAt;
+}
