+++ src/main/java/note/share/service/UserService.java
@@ -1,0 +1,4 @@
+package note.share.service;
+
+public interface UserService {
+}
