+++ src/main/java/note/share/service/impl/UserServiceImpl.java
@@ -1,0 +1,5 @@
+package note.share.service.impl;
+
+public class UserServiceImpl {
+
+}
